refactor: chatbot task types -> chatbot.types.tasks
