+++ src/main/java/li/zhongfu/cs2103/chatbot/types/tasks/Task.java
@@ -1,4 +1,4 @@
-package li.zhongfu.cs2103.chatbot.types;
+package li.zhongfu.cs2103.chatbot.types.tasks;
 
 import java.io.Serializable;
